requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.2.0
future==0.18.2
idna==3.3
idna-ssl==1.1.0
line-bot-sdk==2.2.1
multidict==5.2.0
requests==2.27.1
typing_extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
